+++ src/wsi-anonymizer.h
@@ -13,9 +13,6 @@
 static const char *VENDOR_AND_FORMAT_STRINGS[] = {"Aperio",           "Hamamatsu",     "3DHistech (Mirax)", "Ventana",
                                                   "Philips’ iSyntax", "Philips' TIFF", "Unknown",           "Invalid"};
 
-// TODO: remove this
-extern int8_t check_file_format(const char *filename);
-
 extern struct wsi_data *get_wsi_data(const char *filename);
 
 extern int32_t anonymize_wsi_inplace(const char *filename, const char *new_label_name, bool keep_macro_image,